Tool Development
Within this update I created a Key tool which will be used to open doors within the game. After 5 seconds these doors will close again to promote more use of the tool. I also created a magnet tool, this has a magnetic radius on it and anything with the tag "Magnetic" on it will be pulled toward it this could be utilised in a number of different ways and puzzles. For example pulling a slab over an empty area to create a bridge to cross. As well as the tools functionality I added a key object as well as a magnet object, a door was added with an open and close animation. Currently any magnetic objects are for testing of the functionality and any finalized objects currently are not developed.