Level Design
This update see's the initial development of the different playing areas that the player will explore throughout the game. The first section is a dungeon where the player will navigate dark narrow corridors to find the key to unlock the doors blocking their path and escape through the portal. This leads them to the next area, here the player will explore a more open forest environment in search for a treasure chest. Contained within will be the magnet which they'll use to solve the puzzle in this area. On top of this a selection of music tracks have been lined up and will be used throughout the game, in order to change these tracks the player will collide with an invisible cube outside of each portal. The goal with these tracks is to add variety while maintaining a theme, using the same track throughout would bore the player. Finally I added a skybox which emphasises and backs up the idea that the player is within a simulated world and gives off an essence of unnaturalness to it.