Continued Level Design
The first level has been fully developed and fully playable with all the environment's having their final intended props and effects in place. On top of this the player has been given footsteps for when they walk around. I had noticed within bumpy environment's that sound isn't always consistent this is something that could be tweaked in a later version however.