World 3 Development & Portal Link up
Today I had set up an additional pair of portals linking the 2nd and 3rd world together, some technical issues had drawn out this process unfortunately which lead to more time being spent on this than anticipated. As well as this today I also developed the third world, its theming will be a scrap yard in which the player will be forced to used their magnet tool to navigate and progress through a maze like car grave yard. All that remains for this world is to put in the "Magnetic" interactive objects.