Level 2 & 3 Design Work
All 3 levels are now fully developed and completed including the levels challenege. The second level will see the player using the key to search a small  village in search of a magnet. This magnet will help them to navigate and progress through the 3rd level. The player here will find themselves trapped in a scrap yard. The player must move the cars to uncover the path through to the end and complete the "Test".