Inventory System Implementation & Main Menu
Within this update i have fully implemented inventory functionality. In order to use the items the player must first collide with them to pick them up, these are then added to their inventory. This is represented through the GUI on the top left of the players screen. When the player wants to cycle through the different items they can do so by clicking the E key. On top of this i also created the skeleton for the main menu by selecting "Play game" the player will load in with a fresh save and they begin at the start, they can then save and leave the game and when they return this time they can click load game which will then load the game from where they last saved. If the player clicks load game before saving the button will function as a new game button until they save for the first time. Within the in game UI the player now also has the options of saving the game and exiting which shuts down the program.