Inventory Data Persistence
This update see's the inventory having fully working data persistence. This means when the player saves and exits the game and comes back later to load and continue playing they will have their position and inventory the same as where the player last left off. This update will be the final major update future updates seen past this projects submission will be further polish and improvements that could be made.